requirements.txt
huggingface_hub
python-dotenv
langchain
sentence-transformers
streamlit
langchain-experimental
tabulate